+++ chia-primitives/src/lib.rs
@@ -1,7 +1,8 @@
 pub mod conditions;
 pub mod puzzles;
-pub mod standard_puzzle;
 
 mod derive_synthetic;
+mod primitives;
 
 pub use derive_synthetic::*;
+pub use primitives::*;
